4.context_with_embedding/pyproject.toml
[project]
name = "4-context-with-embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.16",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.30",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.11.0",
]
